+++ packages/scenes/src/components/layout/LazyLoader.tsx
@@ -2,6 +2,8 @@ import React, { ForwardRefExoticComponent, useImperativeHandle, useRef, useState
 import { useEffectOnce } from 'react-use';
 
 import { uniqueId } from 'lodash';
+import { css } from '@emotion/css';
+import { useStyles2 } from '@grafana/ui';
 
 export function useUniqueId(): string {
   const idRefLazy = useRef<string | undefined>(undefined);
@@ -23,8 +25,9 @@ export interface LazyLoaderType extends ForwardRefExoticComponent<Props> {
 }
 
 export const LazyLoader: LazyLoaderType = React.forwardRef<HTMLDivElement, Props>(
-  ({ children, onLoad, onChange, ...rest }, ref) => {
+  ({ children, onLoad, onChange, className, ...rest }, ref) => {
     const id = useUniqueId();
+    const { hideEmpty } = useStyles2(getStyles);
     const [loaded, setLoaded] = useState(false);
     const [isInView, setIsInView] = useState(false);
     const innerRef = useRef<HTMLDivElement>(null);
@@ -57,13 +60,29 @@ export const LazyLoader: LazyLoaderType = React.forwardRef<HTMLDivElement, Props
       };
     });
 
+    // If the element was loaded, we add the `hideEmpty` class to potentially
+    // hide the LazyLoader if it does not have any children. This is the case
+    // when children have the `isHidden` property set. 
+    // We always include the `className` class, as this is coming from the 
+    // caller of the `LazyLoader` component.
+    const classes = `${loaded ? hideEmpty : ''} ${className}`;
     return (
-      <div id={id} ref={innerRef} {...rest}>
+      <div id={id} ref={innerRef} className={classes} {...rest}>
         {loaded && (typeof children === 'function' ? children({ isInView }) : children)}
       </div>
     );
   }
 ) as LazyLoaderType;
+
+function getStyles() {
+  return {
+    hideEmpty: css({
+      '&:empty': {
+        display: 'none',
+      },
+    }),
+  };
+}
 
 LazyLoader.displayName = 'LazyLoader';
 LazyLoader.callbacks = {} as Record<string, (e: IntersectionObserverEntry) => void>;